pyproject.toml
[project]
name = "penguins"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "avr2002", email = "avr13405@gmail.com" }]
requires-python = "~=3.10.0"
dependencies = [
    "comet-ml>=3.50.0",
    "keras>=3.11.2",
    "numpy>=1.26.4",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "sagemaker[local]>=2.250.0",
    "scikit-learn>=1.7.1",
    "tensorflow>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
cdk = ["aws-cdk-lib>=2.208.0", "constructs>=10.4.2"]
notebook = ["ipykernel>=6.29.5", "jupyter>=1.1.1", "rich>=13.9.4"]
static-code-qa = ["pre-commit>=4.2.0", "ruff>=0.12.7"]
stubs = [
    "boto3-stubs[ecr,s3,sagemaker]>=1.39.17",
]
test = ["ipytest>=0.14.2", "pytest>=8.4.1", "pytest-cov>=6.2.1"]

# https://github.com/astral-sh/ruff?tab=readme-ov-file#configuration
[tool.ruff]
line-length = 119
extend-include = ["*.ipynb"]

# https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
extend-select = [
    # pycodestyle errors
    "E",
    # flake8-bugbear
    "B",
    # pylint equivalent rules
    "PL",
    # isort
    "I",
    # pydocstyle for docstrings
    "D",
    # pep8-naming
    "N",
]
ignore = [
    "B904",  # Within an `except` clause, raise exceptions with `raise ... from err` or `raise ... from None` to distinguish them from errors in exception handling
    "E501",  # line-too-long
    "F401",  # unused-import
    "W605",  # invalid-escape-sequence
    "W291",  # trailing-whitespace"
    "UP032", # Use f-string instead of format call
    # pydocstyle - https://docs.astral.sh/ruff/rules/#pydocstyle-d
    "D103", # undocumented-public-function
    "D104", # Missing docstring in public package
    "D107", # Missing docstring in `__init__`
    "D203", # incorrect-blank-line-before-class
    "D100", # undocumented-public-module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D202", # No blank lines allowed after function docstring
    "D205", # 1 blank line required between summary line and description
    "D212", # multi-line-summary-first-line
    "D400", # first line should end with a period
    "D401", # first line should be in imperative mood
    "D415", # First line should end with a period, question mark, or exclamation point
    "N803", # Argument name should be lowercase
    "N806", # Variable name should be lowercase
    # https://docs.astral.sh/ruff/rules/#pylint-pl
    "PLR2004", # Magic value used in comparison, consider replacing `2` with a constant variable
    # Magic value used in comparison, consider replacing `4` with a constant variable -- Is problem for tests
    "PLR2004",
    # `with` statement variable `tmp_dir` overwritten by assignment target, eg: tests/unit_tests/components/metaflow_project/test__metaflow_project_.py:17:9
    "PLW2901",
    "PLC0415", # import should be at the top level of a file
    "PLR0913", # Too many arguments in function definition
]
[tool.ruff.lint.pydocstyle]
convention = "google"
